Add -DSIGNAL_PROVIDED to NaCl build.
This change makes -DSIGNAL_PROVIDED the default for the NaCl
target which will bring the PNaCl build inline with the
x86 and arm builds which define this on the command line.

This define tells newlib to assume that the OS provides the
signal/raise/kill functions and newlib should not include
its own internal versions of these.

Once this lands that command line flags can be removed from
the other builds.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3646
TEST= ./build.sh libs and manual verification
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/22929038