Add NaCl sandboxing to new ARM assembly code
The upstream merge brought in new ARM assembly code.  It needs
adjustments to conform to the NaCl ABI.  The memcpy and strlen
implementations are identical to ones already incorporated into
glibc and for which I'd already done the NaCl SFI changes and
done thorough testing with the glibc test suite for those
functions.  The strcmp implementation is new but the sandboxing
changes there are relatively simple.  I'm fairly confident of
these sandboxing changes (and they certainly pass validation).
Testing the new newlib on ARM has many failures not yet
understood, but there was no testing on ARM done before these
changes so the problems may well be unrelated to this assembly.

BUG= none
TEST= none
R=phosek@chromium.org

Review URL: https://codereview.chromium.org/25112006

Change-Id: Ied8f7c39831ed439e34ba9a3d2402a168c04ce96