Define _POSIX_THREADS under NaCl.
Previously this was not being defined at all, and when it
was defined things would break because the types in
sys/types.h are not compatible with NaCl's pthread.h.

This will allow for the simplification of the python
port in naclports, which relies on _POSIX_THREADS being
defined (if not in features.h then at least in pthread.h).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3551
TEST= trybots + manual testing
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/82523004